Make website header sticky and refine mini cart sliding functionality
To prevent the page from shifting to the right when the mini cart is
opened, a right margin is added to the document body (if a vertical
scrollbar is present). This compensates for the width of the vertical
scrollbar that disappears when the body's overflow property is set to hidden.